incluindo campos de lado das areas fix #83

+++ database/migrations/2023_03_27_194918_add_area_disponivel_expansao.php
@@ -14,7 +14,7 @@ class AddAreaDisponivelExpansao extends Migration
     public function up()
     {
         Schema::table('unidade_produtivas', function (Blueprint $table) {
-            $table->bigInteger('area_disponivel_expansao')->nullable();
+            $table->decimal('area_disponivel_expansao', 8, 2)->nullable();
         });
     }
